Write chain to next index if it doesn't fit in a single index
According to ek-credential spec (v2.5) Section 2.2.1.5.2:
If a concatenated certificate chain does not fit in a single index,
the chain MUST overflow to the next numerically larger index in the
list of NV Indices.

The max_nv_index_size and max_nv_buffer_size are received by
tpm2_get_caps().

Signed-off-by: Min Xu <min.m.xu@intel.com>