Refactor tpm2_provision_ek
Previously ca-cert and ek-cert are both generated in
tpm2_provision_ek(). Based on the vTPM spec, ca-cert shall be generated
on the launch of vtpm-td and it is not destroyed even a DestroyInstance
event is received. While ek-cert shall be generated on the receiving of
CreateInstance event and be destroyed on the receiving of
DestroyInstance event.

This patch refactors tpm2_provision_ek() by getting the ca-cert from
global data.

Signed-off-by: Min Xu <min.m.xu@intel.com>